// src/types/ClassNameButton.ts
export enum ClassNameButton {
  Info = 'button is-info',
  Success = 'button is-success',
  Warning = 'button is-warning',
  Danger = 'button is-danger',
  InfoLight = 'button is-info is-light',
  SuccessLight = 'button is-success is-light',
  WarningLight = 'button is-warning is-light',
  DangerLight = 'button is-danger is-light',
}